2.2 Area of ​​a rhombus

+++ main.py
@@ -1,3 +1,3 @@
-a,b=int(input()),int(input())
+a,b=float(input()),float(input())
 p=(a*b)/2
 print(int(p),round(p,1),p,sep='\n')
